backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "cryptography>=47.0.0",
    "fastapi>=0.136.1",
    "passlib>=1.7.4",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "pyjwt>=2.12.1",
    "python-multipart>=0.0.26",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn>=0.46.0",
]
